requirements.txt
fastapi
uvicorn
ollama
pydantic
mysql-connector-python
